rename `ListItem` to `List.Item` and expose ref
