+++ packages/react/src/components/List/ListItem/ListItem.tsx
@@ -1,15 +1,19 @@
-import type { ComponentPropsWithoutRef } from 'react';
-import React from 'react';
+import type { HTMLAttributes } from 'react';
+import React, { forwardRef } from 'react';
+import cn from 'classnames';
 
 import classes from './ListItem.module.css';
 
-export type ListItemProps = ComponentPropsWithoutRef<'li'>;
+export type ListItemProps = HTMLAttributes<HTMLLIElement>;
 
-export const ListItem = ({ children, ...rest }: ListItemProps) => (
-  <li
-    {...rest}
-    className={classes.listItem}
-  >
-    {children}
-  </li>
+export const ListItem = forwardRef<HTMLLIElement, ListItemProps>(
+  ({ children, ...rest }: ListItemProps, ref) => (
+    <li
+      {...rest}
+      className={cn(classes.listItem, rest.className)}
+      ref={ref}
+    >
+      {children}
+    </li>
+  ),
 );
