+++ packages/react/src/components/List/index.ts
@@ -1,4 +1,17 @@
-export { List } from './List';
-export { ListItem } from './ListItem';
+import { List as ListRoot } from './List';
+import { ListItem } from './ListItem/ListItem';
+
 export type { ListProps } from './List';
-export type { ListItemProps } from './ListItem';
+export type { ListItemProps } from './ListItem/ListItem';
+
+type ListComponent = typeof ListRoot & {
+  Item: typeof ListItem;
+};
+
+const List = ListRoot as ListComponent;
+
+List.Item = ListItem;
+
+List.Item.displayName = 'List.Item';
+
+export { List, ListItem };
